+++ terraform/github/imports.tf
@@ -1,0 +1,63 @@
+# .github
+import {
+  to = module.repository[".github"].github_repository.repo
+  id = ".github"
+}
+
+import {
+  to = module.repository[".github"].github_branch_default.default
+  id = ".github"
+}
+
+import {
+  to = module.repository[".github"].github_repository_ruleset.this["main-protection"]
+  id = ".github:main-protection"
+}
+
+# dotfiles
+import {
+  to = module.repository["dotfiles"].github_repository.repo
+  id = "dotfiles"
+}
+
+import {
+  to = module.repository["dotfiles"].github_branch_default.default
+  id = "dotfiles"
+}
+
+# fried-shrimp
+import {
+  to = module.repository["fried-shrimp"].github_repository.repo
+  id = "fried-shrimp"
+}
+
+import {
+  to = module.repository["fried-shrimp"].github_branch_default.default
+  id = "fried-shrimp"
+}
+
+import {
+  to = module.repository["fried-shrimp"].github_repository_ruleset.this["develop-protection"]
+  id = "fried-shrimp:develop-protection"
+}
+
+import {
+  to = module.repository["fried-shrimp"].github_repository_ruleset.this["main-protection"]
+  id = "fried-shrimp:main-protection"
+}
+
+# scribe
+import {
+  to = module.repository["scribe"].github_repository.repo
+  id = "scribe"
+}
+
+import {
+  to = module.repository["scribe"].github_branch_default.default
+  id = "scribe"
+}
+
+import {
+  to = module.repository["scribe"].github_repository_ruleset.this["main-protection"]
+  id = "scribe:main-protection"
+}